+++ CHANGELOG.md
@@ -46,6 +46,7 @@ Consider migrating to `stream-chat-android-ui-components` or `stream-chat-androi
 ### ⬆️ Improved
 
 ### ✅ Added
+- Added new extension function `ChatClient::keystroke`.
 - Added new extension function `ChatClient::stopTyping`.
 
 ### ⚠️ Changed
@@ -66,6 +67,8 @@ Consider migrating to `stream-chat-android-ui-components` or `stream-chat-androi
 ## stream-chat-android-ui-components
 ### 🐞 Fixed
 - Fixed crash related with creation of MessageOptionsDialogFragment
+- Fixed `GiphyViewHolderStyle#cardBackgroundColor` not getting applied
+
 ### ⬆️ Improved
 
 ### ✅ Added